Split out tree/element module into submodules.
