+++ ftml/src/tree/element/mod.rs
@@ -1,0 +1,29 @@
+/*
+ * tree/element/mod.rs
+ *
+ * ftml - Library to parse Wikidot text
+ * Copyright (C) 2019-2021 Wikijump Team
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+mod collection;
+mod iter_owned;
+mod iter_ref;
+mod object;
+
+pub use self::collection::Elements;
+pub use self::iter_owned::*;
+pub use self::iter_ref::*;
+pub use self::object::Element;
